Refactoring intro -> menu, and player select
